Extract EmptyProposals from InitiativeCards
